Switch code color scheme to light background
Alternating code blocks with black background and the text with white
background creates large contrast boundaries that make it difficult to
read text situated just next to them. They also break flow in "dark
mode" or when colors are inverted on the client. Use the Pygments theme
with white background instead. This is consistent with all major code
sharing platforms, such as Github, and our own code review tooling.